Fix some build failures I introduced.

+++ Pfhorge Source/Data Objects/Marathon Data/PhLight.m
@@ -38,16 +38,16 @@
 
 - (long)exportWithIndex:(NSMutableArray *)index withData:(NSMutableData *)theData mainObjects:(NSSet *)mainObjs
 {
-    long theNumber = [index indexOfObjectIdenticalTo:self];
-    long tmpLong = 0;
-    int i = 0;
+    NSInteger theNumber = [index indexOfObjectIdenticalTo:self];
+    NSInteger tmpLong = 0;
+    NSInteger i = 0;
     
     if (theNumber != NSNotFound)
     {
         return theNumber;
     }
     
-    int myPosition = [index count];
+    NSInteger myPosition = [index count];
     
     [index addObject:self];
     
@@ -91,7 +91,7 @@
     
     if ((int)[index indexOfObjectIdenticalTo:self] != myPosition)
     {
-        NSLog(@"BIG EXPORT ERROR: line %d was not at the end of the index... myPosition = %d", [self getIndex], myPosition);
+		NSLog(@"BIG EXPORT ERROR: line %d was not at the end of the index... myPosition = %ld", [self getIndex], (long)myPosition);
         //return -1;
         //return [index indexOfObjectIdenticalTo:self]
     }
